Refactor code into backend, group and repo
